feat : 입력을 담당하는 Input 과 출력을 담당하는 Output 작성
